mainAgent/sub_agents/marwan/pyproject.toml
[project]
name = "study-agent"
version = "1.0.0"
description = "University Study Assistant using ADK and RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "litellm>=1.0.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
